% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sappui.R
\name{Sappui}
\alias{Sappui}
\title{Sappui}
\usage{
Sappui()
}
\description{
Launch the Sapp user interface in local machine
}
\details{
This function will automatically launch the TSCAN user interface in a web browser. 
The user interface provides many powerful functions which is not available by command line programming.
It also provides a much easier and more convenient way to quickly explore single cell data and construct pseudotime analysis.
The user interface can also be accessed by http://zhiji.shinyapps.io/TSCAN. Neither R nor any packages are required in this online version.
However, it is highly recommended that the user interface be launched locally for faster running speed.
}
\examples{
\dontrun{
   Sappui()
}
}
\author{
Sarwar <sarwar.pustcse1@gmail.com>
}
